Refactor code for checking results in test mode.
Code is split into different methods for each type of test.
It is also mostly relocated to a new class prism.ResultTesting,
since it should not really be tied to the Property object.

The result for testing can now be passed as a Result object,
as well as Double, Integer, etc., with the former being preferred.
This will allow info attached to the Result object to be used
for the correctness, notably precision info.

Also add a new testing regression test - just some constants,
including border cases - infinity, NaN, etc.